# Try faster hashCode