@anthony-swirldslabs : A second revision of the fix on 3/31/2025. This time we always use our own hash code computation code and never rely on JDK directly. Also, the benchmark has been modified to actually test the `hashCode` method specifically. Results for small (16), medium (16K), and large (16M) Bytes objects:

**Before the fix:**
```
Benchmark                            Mode  Cnt          Score          Error  Units
BytesHashCodeBench.hashLargeBytes   thrpt    5         62.785 ±        2.616  ops/s
BytesHashCodeBench.hashMediumBytes  thrpt    5      64916.981 ±     1178.676  ops/s
  BytesHashCodeBench.hashSmallBytes   thrpt    5  156480715.770 ± 10117792.960  ops/s
```

**With the fix:**
```
Benchmark                            Mode  Cnt            Score           Error  Units
BytesHashCodeBench.hashLargeBytes   thrpt    5  17794927753.539 ± 120647815.286  ops/s
BytesHashCodeBench.hashMediumBytes  thrpt    5  17793757339.364 ± 137804766.981  ops/s
BytesHashCodeBench.hashSmallBytes   thrpt    5  17711207197.438 ± 541245113.180  ops/s
```

A 100x improvement even for a small Bytes of size 16. A much larger improvement for larger Bytes objects.



-------------------------
Below is the description from the original revision of the fix from @jasperpotts :
-------------------------


Aims to fix #424 , bellow are results on mac laptop.

<img width="816" alt="hashCodeResults" src="https://github.com/user-attachments/assets/a71310e7-c084-4620-9781-024b91395f05" />